requirements.txt
fastapi
uvicorn
websockets
python-multipart
jinja2